Autonomous mode added
2 autonomous modes.

To be worked on:
- Autonomous selector
- Stages for elevator
- 2nd controller
- reset stages of elevator
- add elevator and grabber to autonomous mode

+++ Goldo/src/org/usfirst/frc/team6651/robot/subsystems/Grabber.java
@@ -19,7 +19,7 @@ public class Grabber extends Subsystem {
         // setDefaultCommand(new Elevator());
     }
 	
-    public void grabber_move(int movement) {
+    public void grabber_move(double movement) {
     		grabberTalon1.set(movement*defaultSpeed);
     }
